+++ src/display.c
@@ -43,6 +43,78 @@
 #include "integrator.h"
 #define MAX(a, b) ((a) < (b) ? (b) : (a))       ///< Returns the maximum of a and b
 
+
+static void reb_display_set_default_view(struct reb_simulation* const r, struct reb_display_settings* s){
+    float  scale = 0.;
+    // Need a scale for visualization
+    if (r->root_size==-1){  
+        scale = 0.;
+        const struct reb_particle* p = r->particles;
+        for (unsigned int i=0;i<r->N-r->N_var;i++){
+            const double _r = sqrt(p[i].x*p[i].x+p[i].y*p[i].y+p[i].z*p[i].z);
+            scale = MAX(scale, _r);
+        }
+        if(scale==0.){
+            scale = 1.;
+        }
+        scale *= 1.1;
+    }else{
+        scale = r->boxsize_max/2.;
+    }
+    
+    struct reb_mat4df oldview = s->view;
+    s->view = reb_mat4df_scale(reb_mat4df_identity(), 1./scale, 1./scale, 1./scale);
+    if (oldview.m[1]==0. && oldview.m[2]==0. && oldview.m[4]==0. && oldview.m[6]==0.){
+        struct reb_rotation rotation = {
+            .ix = 1./sqrt(2.),
+            .iy = 0.,
+            .iz = 0.,
+            .r = 1./sqrt(2.),
+        };
+        s->view = reb_mat4df_multiply(reb_rotation_to_mat4df(rotation), s->view);
+    }else if (oldview.m[1]==0. && oldview.m[2]==0. && oldview.m[4]==0. && oldview.m[5]==0.){
+        struct reb_rotation rotation = {
+            .ix = 0.,
+            .iy = -1./sqrt(2.),
+            .iz = 0.,
+            .r = 1./sqrt(2.),
+        };
+        s->view = reb_mat4df_multiply(reb_rotation_to_mat4df(rotation), s->view);
+    }
+}
+
+void reb_display_settings_init(struct reb_simulation*r, struct reb_display_settings* s){
+    if (r->max_radius0 > 0.0){
+        s->spheres       = 1; 
+    }else{
+        s->spheres       = 0; 
+    }
+    s->pause             = 0; 
+    s->multisample       = 1; 
+    if (r->integrator==REB_INTEGRATOR_WHFAST){
+        s->wire          = 1; 
+    }else{
+        s->wire          = 0; 
+    }
+    s->past              = 0; 
+    s->past_N            = 64;
+    s->onscreentext      = 1; 
+    s->ghostboxes        = 0; 
+    s->reference         = -1;
+    s->view.m[1]=1; // this will make set_default_view show the xy plane
+    reb_display_set_default_view(r, s);
+}
+
+void reb_simulation_add_display_settings(struct reb_simulation*r){
+    if (r->display_settings){
+        reb_simulation_error(r,"Simulation already has display settings.");
+        return;
+    }
+    r->display_settings = calloc(1,sizeof(struct reb_display_settings));
+    reb_display_settings_init(r, r->display_settings);
+}
+
+
 #ifdef OPENGL
 #include "simplefont.h"
 
@@ -117,7 +189,6 @@ EM_JS(void, reb_overlay_hide, (int hide), {
 #endif
 
 void reb_render_frame(void* p);
-static void reb_display_set_default_scale(struct reb_simulation* const r);
                 
 static const char* onscreenhelp[] = { 
                 "REBOUND mouse and keyboard commands",
@@ -150,58 +221,6 @@ static const char* onscreenhelp[] = {
                 "----------------------------------------------------"
 };
 
-
-static void matscale(float mat[16], float s){
-    mat[0] = s; mat[1] = 0.; mat[2] = 0.; mat[3] = 0.;
-    mat[4] = 0.; mat[5] = s; mat[6] = 0.; mat[7] = 0.;
-    mat[8] = 0.; mat[9] = 0.; mat[10] = s; mat[11] = 0.;
-    mat[12] = 0.; mat[13] = 0.; mat[14] = 0.; mat[15] = 1.;
-}
-
-static void matscale3(float mat[16], float s[3]){
-    mat[0] = s[0]; mat[1] = 0.; mat[2] = 0.; mat[3] = 0.;
-    mat[4] = 0.; mat[5] = s[1]; mat[6] = 0.; mat[7] = 0.;
-    mat[8] = 0.; mat[9] = 0.; mat[10] = s[2]; mat[11] = 0.;
-    mat[12] = 0.; mat[13] = 0.; mat[14] = 0.; mat[15] = 1.;
-}
-
-static void mattranslate(float mat[16], float x, float y, float z){
-    mat[0] = 1.; mat[1] = 0.; mat[2] = 0.; mat[3] = x; 
-    mat[4] = 0.; mat[5] = 1.; mat[6] = 0.; mat[7] = y; 
-    mat[8] = 0.; mat[9] = 0.; mat[10] = 1.; mat[11] = z;
-    mat[12] = 0.; mat[13] = 0.; mat[14] = 0.; mat[15] = 1.;
-}
-
-static void matortho(float mat[16], float l, float r, float b, float t, float n, float f) {
-    mat[0] = 2.f/(r-l); mat[1] = 0.; mat[2] = 0.; mat[3] = -(r+l)/(r-l);
-    mat[4] = 0.; mat[5] = 2.f/(t-b); mat[6] = 0.; mat[7] = -(t+b)/(t-b);
-    mat[8] = 0.; mat[9] = 0.; mat[10] = -2.f/(f-n); mat[11] = -(f+n)/(f-n);
-    mat[12] = 0.; mat[13] = 0.; mat[14] = 0.; mat[15] = 1.f;
-}
-static void rotation2mat(struct reb_rotation A, float mat[16]){
-    float xx = A.ix * A.ix; float xy = A.ix * A.iy; float xz = A.ix * A.iz;
-    float xw = A.ix * A.r; float yy = A.iy * A.iy; float yz = A.iy * A.iz;
-    float yw = A.iy * A.r; float zz = A.iz * A.iz; float zw = A.iz * A.r;
-    mat[0] = 1.-2.*(yy+zz);
-    mat[1] =    2.*(xy-zw);
-    mat[2] =    2.*(xz+yw);
-    mat[4] =    2.*(xy+zw);
-    mat[5] = 1.-2.*(xx+zz);
-    mat[6] =    2.*(yz-xw);
-    mat[8] =    2.*(xz-yw);
-    mat[9] =    2.*(yz+xw);
-    mat[10]= 1.-2.*(xx+yy);
-    mat[3] = mat[7] = mat[11] = mat[12] = mat[13] = mat[14] = 0; mat[15]= 1;
-}
-
-static void matmult(float A[16], float B[16], float C[16]) {
-    for(int i=0;i<4;i++){
-    for(int j=0;j<4;j++){
-        C[i+4*j] = 0.;
-    for(int k=0;k<4;k++){
-        C[i+4*j] += A[k+4*j]*B[i+4*k];
-    }}}
-}
 
 static int convertLine(const char* in, float* out){
     int j = 0;
@@ -303,7 +322,8 @@ static void reb_display_scroll(GLFWwindow* window, double xoffset, double yoffse
         printf("Error accessing data in reb_display_scroll\n");
         return;
     }
-    data->scale *= 1.-yoffset/100.;
+    float scale = 1.-yoffset/100.;
+    data->s.view = reb_mat4df_scale(data->s.view, scale, scale, scale);
 }
 static void reb_display_mouse_button(GLFWwindow* window, int button, int action, int mods){
     struct reb_display_data* data = glfwGetWindowUserPointer(window);
@@ -340,7 +360,6 @@ static void reb_display_cursor(GLFWwindow* window, double x, double y){
         // New drag event
         data->mouse_x = x;
         data->mouse_y = y;
-        data->mouse_scale = data->scale;
         return;
     }
     if (data->mouse_action==GLFW_PRESS){
@@ -348,32 +367,10 @@ static void reb_display_cursor(GLFWwindow* window, double x, double y){
             // Drag 
             float dx = 3.*(x-data->mouse_x)/width;
             float dy = 3.*(y-data->mouse_y)/height;
-            data->mouse_x = x;
-            data->mouse_y = y;
-
-            struct reb_rotation inv = reb_rotation_conjugate(data->view);
-            struct reb_vec3d up = {.x=0.,.y=1.,.z=0.};
-            struct reb_vec3d right = {.x=1.,.y=0.,.z=0.};
-            struct reb_vec3d inv_right = reb_vec3d_rotate(right, inv);
-            struct reb_vec3d inv_up = reb_vec3d_rotate(up, inv);
-
-            float sin_dy = sin(dy);
-            struct reb_rotation rot_dy;
-            rot_dy.ix    = inv_right.x*sin_dy;
-            rot_dy.iy    = inv_right.y*sin_dy;
-            rot_dy.iz    = inv_right.z*sin_dy;
-            rot_dy.r    = cos(dy);
-            rot_dy = reb_rotation_normalize( rot_dy );
-            data->view = reb_rotation_mul(data->view,rot_dy);
-            
-            float sin_dx = sin(dx);
-            struct reb_rotation rot_dx;
-            rot_dx.ix    = inv_up.x*sin_dx;
-            rot_dx.iy    = inv_up.y*sin_dx;
-            rot_dx.iz    = inv_up.z*sin_dx;
-            rot_dx.r    = cos(dx);
-            rot_dx = reb_rotation_normalize(rot_dx);
-            data->view = reb_rotation_mul(data->view,rot_dx);
+            struct reb_rotation rot_dy = {.ix=sin(dy), .r=cos(dy)};
+            struct reb_rotation rot_dx = {.iy=sin(dx), .r=cos(dx)};
+            data->s.view = reb_mat4df_multiply(reb_rotation_to_mat4df(rot_dy), data->s.view);
+            data->s.view = reb_mat4df_multiply(reb_rotation_to_mat4df(rot_dx), data->s.view);
         }else{
             // Zoom
             float ix = data->mouse_x/width-0.5;
@@ -382,9 +379,11 @@ static void reb_display_cursor(GLFWwindow* window, double x, double y){
             float nx = x/width-0.5;
             float ny = y/height-0.5;
             float nr = sqrt(nx*nx + ny*ny);
-            data->scale = data->mouse_scale*ir/nr;
-
+            data->s.view = reb_mat4df_scale(data->s.view, nr/ir, nr/ir, nr/ir);
         }
+        data->mouse_x = x;
+        data->mouse_y = y;
+        return;
     }
 }
                 
@@ -403,7 +402,7 @@ static void reb_display_clear_particle_data(struct reb_display_data* data){
         }
         glBindBuffer(GL_ARRAY_BUFFER, data->particle_buffer);
         for (int i=0; i<N_hist; i++){
-            glBufferSubData(GL_ARRAY_BUFFER, i*N_real*sizeof(struct reb_vec4d_opengl), N_real*sizeof(struct reb_vec4d_opengl), data->particle_data);
+            glBufferSubData(GL_ARRAY_BUFFER, i*N_real*sizeof(struct reb_vec4df), N_real*sizeof(struct reb_vec4df), data->particle_data);
         }
     }
     if (data->orbit_data){
@@ -443,7 +442,7 @@ void reb_display_keyboard(GLFWwindow* window, int key, int scancode, int action,
     if (action==GLFW_PRESS){
         switch(key){
             case 'H':
-                data->onscreenhelp = !data->onscreenhelp;
+                data->s.onscreenhelp = !data->s.onscreenhelp;
                 break;
             case 'Q':
                 data->r->status = REB_STATUS_USER;
@@ -458,70 +457,54 @@ void reb_display_keyboard(GLFWwindow* window, int key, int scancode, int action,
                 }
                 break;
             case 'S':
-                data->spheres = (data->spheres+1)%4;
+                data->s.spheres = (data->s.spheres+1)%4;
                 break;
             case 'G':
-                data->ghostboxes = !data->ghostboxes;
+                data->s.ghostboxes = !data->s.ghostboxes;
                 break;
             case 'M':
-                data->multisample = !data->multisample;
-                if (data->multisample){
+                data->s.multisample = !data->s.multisample;
+                if (data->s.multisample){
                     glEnable(GL_MULTISAMPLE); 
                 }else{
                     glDisable(GL_MULTISAMPLE); 
                 }
                 break;
             case 'R':
-                if (data->view.r ==1.){
-                    data->view.ix = 1./sqrt(2.);
-                    data->view.iy = 0.;
-                    data->view.iz = 0.;
-                    data->view.r = 1./sqrt(2.);
-                }else if (data->view.ix == 1./sqrt(2.)){
-                    data->view.ix = 0.;
-                    data->view.iy = -1./sqrt(2.);
-                    data->view.iz = 0.;
-                    data->view.r = 1./sqrt(2.);
-                }else{
-                    data->view.ix = 0.;
-                    data->view.iy = 0.;
-                    data->view.iz = 0.;
-                    data->view.r = 1.;
-                }
-                data->reference     = -1;
-                reb_display_set_default_scale(data->r);
+                data->s.reference     = -1;
+                reb_display_set_default_view(data->r, &data->s);
                 break;
             case 'D':
-                data->pause = !data->pause;
+                data->s.pause = !data->s.pause;
                 break;
             case 'W':
-                data->wire = (data->wire+1)%2;
+                data->s.wire = (data->s.wire+1)%2;
                 break;
             case 'C':
                 reb_display_clear_particle_data(data);
                 break;
             case 'P':
-                data->past = (data->past+1)%2;
+                data->s.past = (data->s.past+1)%2;
                 break;
             case 'I':
-                data->past_N *= 2;
+                data->s.past_N *= 2;
                 break;
             case 'O':
-                data->past_current_index = data->past_current_index % MAX(2, data->past_N/2) ;
-                data->past_N = MAX(2, data->past_N/2) ;
+                data->past_current_index = data->past_current_index % MAX(2, data->s.past_N/2) ;
+                data->s.past_N = MAX(2, data->s.past_N/2) ;
                 break;
             case 'T':
-                data->onscreentext = !data->onscreentext;
+                data->s.onscreentext = !data->s.onscreentext;
                 break;
             case 'X': 
                 if (mods!=GLFW_MOD_SHIFT){
-                    data->reference++;
-                    if (data->reference>data->r->N) data->reference = -1;
-                    printf("Reference particle: %d.\n",data->reference);
+                    data->s.reference++;
+                    if (data->s.reference>data->r->N) data->s.reference = -1;
+                    printf("Reference particle: %d.\n",data->s.reference);
                 }else{
-                    data->reference--;
-                    if (data->reference<-1) data->reference = data->r->N-1;
-                    printf("Reference particle: %d.\n",data->reference);
+                    data->s.reference--;
+                    if (data->s.reference<-1) data->s.reference = data->r->N-1;
+                    printf("Reference particle: %d.\n",data->s.reference);
                 }
                 break;
             case 264: // arrow down
@@ -588,27 +571,35 @@ void reb_render_frame(void* p){
         data->need_copy = 0;
         pthread_mutex_unlock(&(data->mutex));  
     }
+    
+    if (r_copy->display_settings){
+        // User provided settings server-side. Will overwrite our own.
+        data->s = *r_copy->display_settings;
+        if (data->s.past_N <2){ 
+            data->s.past_N = 2;
+        }
+    }
 
     // prepare data (incl orbit calculation)
     const int N_real = r_copy->N - r_copy->N_var;
-
-    if (N_real > data->N_allocated || data->past_N != data->past_N_allocated){
-        data->particle_data = realloc(data->particle_data, N_real*sizeof(struct reb_vec4d_opengl));
+    
+    if (N_real > data->N_allocated || data->s.past_N != data->past_N_allocated){
+        data->particle_data = realloc(data->particle_data, N_real*sizeof(struct reb_vec4df));
         data->orbit_data = realloc(data->orbit_data, N_real*sizeof(struct reb_orbit_opengl));
         
         // Resize memory if needed
         glBindBuffer(GL_ARRAY_BUFFER, data->particle_buffer);
-        glBufferData(GL_ARRAY_BUFFER, data->past_N*N_real*sizeof(struct reb_vec4d_opengl), NULL, GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, data->s.past_N*N_real*sizeof(struct reb_vec4df), NULL, GL_STATIC_DRAW);
         glBindBuffer(GL_ARRAY_BUFFER, data->particle_buffer_current);
-        glBufferData(GL_ARRAY_BUFFER, N_real*sizeof(struct reb_vec4d_opengl), NULL, GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, N_real*sizeof(struct reb_vec4df), NULL, GL_STATIC_DRAW);
         
         glBindBuffer(GL_ARRAY_BUFFER, data->orbit_buffer);
-        glBufferData(GL_ARRAY_BUFFER, data->past_N*N_real*sizeof(struct reb_orbit_opengl), NULL, GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, data->s.past_N*N_real*sizeof(struct reb_orbit_opengl), NULL, GL_STATIC_DRAW);
         glBindBuffer(GL_ARRAY_BUFFER, data->orbit_buffer_current);
         glBufferData(GL_ARRAY_BUFFER, N_real*sizeof(struct reb_orbit_opengl), NULL, GL_STATIC_DRAW);
         
         data->N_allocated = N_real;
-        data->past_N_allocated = data->past_N;
+        data->past_N_allocated = data->s.past_N;
                 
         reb_display_clear_particle_data(data);
     }
@@ -633,7 +624,7 @@ void reb_render_frame(void* p){
         data->past_current_index = (data->past_current_index+1) % data->past_N_allocated;
     }
 
-    if (data->wire && N_real>1){
+    if (data->s.wire && N_real>1){
         struct reb_particle com = r_copy->particles[0];
         for (unsigned int i=1;i<N_real;i++){
             struct reb_particle p = r_copy->particles[i];
@@ -653,15 +644,15 @@ void reb_render_frame(void* p){
     if (N_real>0){
         // Fill memory (but not resize)
         glBindBuffer(GL_ARRAY_BUFFER, data->particle_buffer);
-        glBufferSubData(GL_ARRAY_BUFFER, data->past_current_index*N_real*sizeof(struct reb_vec4d_opengl), N_real*sizeof(struct reb_vec4d_opengl), data->particle_data);
-        if (data->spheres==1 || data->spheres==2){
+        glBufferSubData(GL_ARRAY_BUFFER, data->past_current_index*N_real*sizeof(struct reb_vec4df), N_real*sizeof(struct reb_vec4df), data->particle_data);
+        if (data->s.spheres==1 || data->s.spheres==2){
             glBindBuffer(GL_ARRAY_BUFFER, data->particle_buffer_current);
-            glBufferSubData(GL_ARRAY_BUFFER, 0, N_real*sizeof(struct reb_vec4d_opengl), data->particle_data);
+            glBufferSubData(GL_ARRAY_BUFFER, 0, N_real*sizeof(struct reb_vec4df), data->particle_data);
         }
             
         glBindBuffer(GL_ARRAY_BUFFER, data->orbit_buffer);
         glBufferSubData(GL_ARRAY_BUFFER, data->past_current_index*(N_real-1)*sizeof(struct reb_orbit_opengl), (N_real-1)*sizeof(struct reb_orbit_opengl), data->orbit_data);
-        if (data->wire){
+        if (data->s.wire){
             glBindBuffer(GL_ARRAY_BUFFER, data->orbit_buffer_current);
             glBufferSubData(GL_ARRAY_BUFFER, 0, (N_real-1)*sizeof(struct reb_orbit_opengl), data->orbit_data);
         }
@@ -677,38 +668,26 @@ void reb_render_frame(void* p){
 #endif
    
     // Precalculate matricies 
-    float projection[16];
-    matortho( projection,
-            -1.6*ratio*data->scale, 1.6*ratio*data->scale, 
-            -1.6*data->scale,1.6*data->scale,
-            -2.5*data->scale,2.5*data->scale);
-    float view[16];
-    float tmp1[16];
-    float tmp2[16];
-    float tmp3[16];
-    if (data->reference>=0){
-        struct reb_particle p = data->r_copy->particles[data->reference];
-        mattranslate(tmp2,-p.x,-p.y,-p.z);
-        rotation2mat(data->view,tmp1);
-        matmult(tmp1,tmp2,view);
-    }else{
-        rotation2mat(data->view,view);
+    struct reb_mat4df projection = reb_mat4df_ortho( -1.6*ratio, 1.6*ratio, -1.6,1.6, -2.5,2.5);
+    struct reb_mat4df view = data->s.view;
+    if (data->s.reference>=0){
+        struct reb_particle p = data->r_copy->particles[data->s.reference];
+        view = reb_mat4df_translate(view, -p.x, -p.y, -p.z);
     }
     
-    for (int i=-data->ghostboxes*data->r_copy->N_ghost_x;i<=data->ghostboxes*data->r_copy->N_ghost_x;i++){
-    for (int j=-data->ghostboxes*data->r_copy->N_ghost_y;j<=data->ghostboxes*data->r_copy->N_ghost_y;j++){
-    for (int k=-data->ghostboxes*data->r_copy->N_ghost_z;k<=data->ghostboxes*data->r_copy->N_ghost_z;k++){
+    for (int i=-data->s.ghostboxes*data->r_copy->N_ghost_x;i<=data->s.ghostboxes*data->r_copy->N_ghost_x;i++){
+    for (int j=-data->s.ghostboxes*data->r_copy->N_ghost_y;j<=data->s.ghostboxes*data->r_copy->N_ghost_y;j++){
+    for (int k=-data->s.ghostboxes*data->r_copy->N_ghost_z;k<=data->s.ghostboxes*data->r_copy->N_ghost_z;k++){
         struct reb_vec6d gb = reb_boundary_get_ghostbox(data->r_copy, i,j,k);
+        struct reb_mat4df model = reb_mat4df_translate(reb_mat4df_identity(), gb.x, gb.y, gb.z);
         { // Particles
-            mattranslate(tmp2,gb.x,gb.y,gb.z);
-            matmult(view,tmp2,tmp1);
-            matmult(projection,tmp1,tmp2);
-            if(data->spheres==1||data->spheres==2){
+            struct reb_mat4df mvp = reb_mat4df_multiply(projection, reb_mat4df_multiply(view, model));
+            if(data->s.spheres==1||data->s.spheres==2){
                 // Solid Spheres
                 glEnable(GL_DEPTH_TEST);
                 glUseProgram(data->shader_sphere.program);
-                glUniformMatrix4fv(data->shader_sphere.mvp_location, 1, GL_TRUE, (GLfloat*) tmp2);
-                if (data->past==1){
+                glUniformMatrix4fv(data->shader_sphere.mvp_location, 1, GL_TRUE, (GLfloat*) mvp.m);
+                if (data->s.past==1){
                     glBindVertexArray(data->shader_sphere.particle_vao);
                     reb_glDrawArraysInstanced(GL_TRIANGLE_STRIP, 0, 800, N_real*data->past_N_allocated);
                 }else{
@@ -719,13 +698,13 @@ void reb_render_frame(void* p){
                 glDisable(GL_DEPTH_TEST);
             }
 
-            if(data->spheres%2==0){
+            if(data->s.spheres%2==0){
                 glUseProgram(data->shader_point.program);
                 glBindVertexArray(data->shader_point.particle_vao);
-                glUniformMatrix4fv(data->shader_point.mvp_location, 1, GL_TRUE, (GLfloat*) tmp2);
+                glUniformMatrix4fv(data->shader_point.mvp_location, 1, GL_TRUE, (GLfloat*) mvp.m);
                 glUniform1i(data->shader_point.past_N_location, data->past_N_allocated);
                 int current_vertex = (data->past_current_index-1+data->past_N_allocated)%data->past_N_allocated;
-                if (data->past==1){
+                if (data->s.past==1){
                     glUniform4f(data->shader_point.color_location, 1.,1.,1.,0.8);
                     glUniform1i(data->shader_point.N_real_location, N_real);
                     glUniform1i(data->shader_point.current_vertex_location, current_vertex);
@@ -737,12 +716,12 @@ void reb_render_frame(void* p){
                 glDrawArrays(GL_POINTS, N_real*((current_vertex+1)%data->past_N_allocated), N_real);
                 glBindVertexArray(0);
             }
-            if (data->wire==1){
+            if (data->s.wire==1){
                 // Orbits
                 glUseProgram(data->shader_orbit.program);
-                glUniformMatrix4fv(data->shader_orbit.mvp_location, 1, GL_TRUE, (GLfloat*) tmp2);
+                glUniformMatrix4fv(data->shader_orbit.mvp_location, 1, GL_TRUE, (GLfloat*) mvp.m);
                 glUniform1i(data->shader_orbit.past_N_location, data->past_N_allocated);
-                if (data->past==1){
+                if (data->s.past==1){
                     glBindVertexArray(data->shader_orbit.particle_vao);
                     glUniform1i(data->shader_orbit.N_real_location, N_real-1);
                     int current_vertex = (data->past_current_index-1+data->past_N_allocated)%data->past_N_allocated;
@@ -760,18 +739,15 @@ void reb_render_frame(void* p){
             glUseProgram(data->shader_box.program);
             if (data->r_copy->boundary == REB_BOUNDARY_NONE){
                 glBindVertexArray(data->shader_box.cross_vao);
-                matscale(tmp1,data->scale);
+                struct reb_vec3df scale = reb_mat4df_get_scale(view); // Extract scale from view matrix so it can be undone
+                model = reb_mat4df_scale(model, 1./scale.x, 1./scale.y, 1./scale.z);
             }else{
                 glBindVertexArray(data->shader_box.box_vao);
-                float boxsize[3] = {data->r_copy->boxsize.x/2., data->r_copy->boxsize.y/2., data->r_copy->boxsize.z/2.};
-                matscale3(tmp1,boxsize);
+                model = reb_mat4df_scale(model, data->r_copy->boxsize.x/2., data->r_copy->boxsize.y/2., data->r_copy->boxsize.z/2.);
             }
+            struct reb_mat4df mvp = reb_mat4df_multiply(projection, reb_mat4df_multiply(view, model));
+            glUniformMatrix4fv(data->shader_box.mvp_location, 1, GL_TRUE, (GLfloat*) mvp.m);
             glUniform4f(data->shader_box.color_location, 1.,0.,0.,1.);
-            mattranslate(tmp2,gb.x,gb.y,gb.z);
-            matmult(tmp2,tmp1,tmp3);
-            matmult(view,tmp3,tmp1);
-            matmult(projection,tmp1,tmp2);
-            glUniformMatrix4fv(data->shader_box.mvp_location, 1, GL_TRUE, (GLfloat*) tmp2);
             if (data->r_copy->boundary == REB_BOUNDARY_NONE){
                 glDrawArrays(GL_LINES, 0, 6);
             }else{
@@ -781,7 +757,7 @@ void reb_render_frame(void* p){
         }
     }}}
 #ifndef __EMSCRIPTEN__
-    if (data->onscreentext){ // On screen text
+    if (data->s.onscreentext){ // On screen text
         glUseProgram(data->shader_simplefont.program);
         glBindVertexArray(data->shader_simplefont.vao);
         glBindTexture(GL_TEXTURE_2D,data->shader_simplefont.texture);
@@ -852,7 +828,7 @@ void reb_render_frame(void* p){
         glBindVertexArray(0);
         glBindTexture(GL_TEXTURE_2D,0);
     }
-    if (data->onscreenhelp){ // On screen help
+    if (data->s.onscreenhelp){ // On screen help
         glUseProgram(data->shader_simplefont.program);
         glBindVertexArray(data->shader_simplefont.vao);
         glBindTexture(GL_TEXTURE_2D,data->shader_simplefont.texture);
@@ -883,12 +859,12 @@ EM_BOOL reb_render_frame_emscripten(double time, void* p){
     if (!data){
         return EM_TRUE;
     }
-    if (data->pause){
+    if (data->s.pause){
         return EM_TRUE;
     }
     reb_render_frame(data);
-    reb_overlay_hide(!data->onscreentext);
-    if (data->onscreentext){ 
+    reb_overlay_hide(!data->s.onscreentext);
+    if (data->s.onscreentext){ 
         char str[10240] = "\0";
         char line[1024];
         sprintf(line,"<div class=\"reboundlogo\"></div>REBOUND v%s<br />",reb_version_str);
@@ -926,8 +902,8 @@ EM_BOOL reb_render_frame_emscripten(double time, void* p){
             reb_overlay_update(str, 10);
         }
     }
-    data->onscreenhelp = reb_overlay_help_show(data->onscreenhelp);
-    if (data->onscreenhelp){ 
+    data->s.onscreenhelp = reb_overlay_help_show(data->s.onscreenhelp);
+    if (data->s.onscreenhelp){ 
         char str[10240] = "\0";
         for (int i=0;i<sizeof(onscreenhelp)/sizeof(onscreenhelp[0]);i++){
             strlcat(str, onscreenhelp[i], 10240);
@@ -938,6 +914,7 @@ EM_BOOL reb_render_frame_emscripten(double time, void* p){
     return EM_TRUE;
 }
 #endif
+
 
 void reb_display_init(struct reb_simulation * const r){
     struct reb_display_data* data = r->display_data;
@@ -972,34 +949,16 @@ void reb_display_init(struct reb_simulation * const r){
     glfwSetWindowUserPointer(window,data); 
 
     // Default parameters
-    reb_display_set_default_scale(r);
+    reb_display_settings_init(r, &r->display_data->s);
     { // Check if we have a retina display
         int wwidth, wheight, fwidth, fheight;
         glfwGetWindowSize(window, &wwidth, &wheight);
         glfwGetFramebufferSize(window, &fwidth, &fheight);
         data->retina = (double)fwidth/(double)wwidth;
     }
-    if (data->r->max_radius0 > 0.0){
-        data->spheres       = 1; 
-    }else{
-        data->spheres       = 0; 
-    }
-    data->pause             = 0; 
-    data->multisample       = 1; 
-    if (data->r->integrator==REB_INTEGRATOR_WHFAST){
-        data->wire          = 1; 
-    }else{
-        data->wire          = 0; 
-    }
-    data->past              = 0; 
+    data->window            = window;
     data->past_current_index= 0;
     data->past_N_allocated  = 0;
-    data->past_N            = 64;
-    data->onscreentext      = 1; 
-    data->ghostboxes        = 0; 
-    data->reference         = -1;
-    data->view.r            = 1.;
-    data->window            = window;
 
     glfwSetKeyCallback(window,reb_display_keyboard);
     glfwSetScrollCallback(window,reb_display_scroll);
@@ -1523,7 +1482,7 @@ void reb_display_init(struct reb_simulation * const r){
 
     while(!glfwWindowShouldClose(window) && r->status<0){
         double t0 = glfwGetTime();
-        if (!data->pause){
+        if (!data->s.pause){
             reb_render_frame(data);
         }
         while (glfwGetTime()-t0 < 1.0/120.) { // Maxframerate 120Hz
@@ -1544,24 +1503,6 @@ void reb_display_init(struct reb_simulation * const r){
     }
     glfwTerminate();
 #endif
-}
-
-static void reb_display_set_default_scale(struct reb_simulation* const r){
-    // Need a scale for visualization
-    if (r->root_size==-1){  
-        r->display_data->scale = 0.;
-        const struct reb_particle* p = r->particles;
-        for (unsigned int i=0;i<r->N-r->N_var;i++){
-            const double _r = sqrt(p[i].x*p[i].x+p[i].y*p[i].y+p[i].z*p[i].z);
-            r->display_data->scale = MAX(r->display_data->scale, _r);
-        }
-        if(r->display_data->scale==0.){
-            r->display_data->scale = 1.;
-        }
-        r->display_data->scale *= 1.1;
-    }else{
-        r->display_data->scale = r->boxsize_max/2.;
-    }
 }
 
 #endif // OPENGL