+++ src/rebound.c
@@ -66,7 +66,7 @@ void usleep(__int64 usec);
 const int reb_max_messages_length = 1024;   // needs to be constant expression for array size
 const int reb_N_max_messages = 10;
 const char* reb_build_str = __DATE__ " " __TIME__;  // Date and time build string. 
-const char* reb_version_str = "4.1.1";         // **VERSIONLINE** This line gets updated automatically. Do not edit manually.
+const char* reb_version_str = "4.2.0";         // **VERSIONLINE** This line gets updated automatically. Do not edit manually.
 const char* reb_githash_str = STRINGIFY(GITHASH);             // This line gets updated automatically. Do not edit manually.
 
 static int reb_simulation_error_message_waiting(struct reb_simulation* const r);
@@ -310,6 +310,9 @@ void reb_simulation_free_pointers(struct reb_simulation* const r){
     if (r->simulationarchive_filename){
         free(r->simulationarchive_filename);
     }
+    if(r->display_settings){
+        free(r->display_settings);
+    }
 #ifdef OPENGL
     if(r->display_data){
         // Waiting for visualization to shut down.
@@ -521,7 +524,9 @@ void reb_simulation_init(struct reb_simulation* r){
     r->output_timing_last   = -1;
     r->save_messages = 0;
     r->track_energy_offset = 0;
+    r->server_data = NULL;
     r->display_data = NULL;
+    r->display_settings = NULL;
     r->walltime = 0;
 
     r->minimum_collision_velocity = 0;