+++ src/rebound.h
@@ -106,6 +106,7 @@ struct reb_display_data;
 struct reb_server_data;
 struct reb_treecell;
 struct reb_variational_configuration;
+struct reb_display_settings;
 
 // Particle structure
 struct reb_particle {
@@ -145,6 +146,12 @@ struct reb_vec3d {
     double y;
     double z;
 };
+
+// Generic 4d matrix (single precision)
+struct reb_mat4df {
+    float m[16];
+};
+
 
 // Generic 6d vector
 struct reb_vec6d{
@@ -472,6 +479,7 @@ struct reb_simulation {
     double exit_max_distance;       // Exit simulation if a particle is this far away from the origin.
     double exit_min_distance;       // Exit simulation if two particles come this close to each other.
     double usleep;                  // Artificially slow down simulations by this many microseconds each timestep.
+    struct reb_display_settings* display_settings;// Optional. Will overwrite settings for visualization. If NULL, UI will determine settings.
     struct reb_display_data* display_data; // Datastructure stores visualization related data. Does not have to be modified by the user. 
     struct reb_server_data* server_data; // Datastructure stores server related data. Does not have to be modified by the user. 
     int track_energy_offset;        // 0 (default): do not track energy offset due to merging/lost particles, 1: track offset
@@ -1035,6 +1043,7 @@ DLLEXPORT struct reb_rotation reb_rotation_init_angle_axis(const double angle, s
 DLLEXPORT struct reb_rotation reb_rotation_init_from_to(struct reb_vec3d from, struct reb_vec3d to);
 DLLEXPORT struct reb_rotation reb_rotation_init_orbit(const double Omega, const double inc, const double omega);
 DLLEXPORT struct reb_rotation reb_rotation_init_to_new_axes(struct reb_vec3d newz, struct reb_vec3d newx);
+DLLEXPORT struct reb_rotation reb_rotation_slerp(struct reb_rotation q1, struct reb_rotation q2, double t);
 
 // transformations to/from vec3d
 DLLEXPORT struct reb_vec3d reb_tools_spherical_to_xyz(const double mag, const double theta, const double phi);
@@ -1071,7 +1080,11 @@ struct reb_orbit_opengl {
     float omega, Omega, inc;
 };
 
-struct reb_vec4d_opengl {
+struct reb_vec3df {
+    float x,y,z;
+};
+
+struct reb_vec4df {
     float x,y,z,r;
 };
 
@@ -1093,14 +1106,27 @@ struct reb_server_data {
 #endif // SERVER
 };
 
+struct reb_display_settings {
+    struct reb_mat4df view;
+    int spheres;                    // Switches between point sprite and real spheres.
+    int pause;                      // Pauses visualization, but keep simulation running
+    int wire;                       // Shows/hides orbit wires.
+    int past;                       // Shows/hides past particle positions.
+    unsigned int past_N;            // Number of past particle positions.
+    int onscreentext;               // Shows/hides onscreen text.
+    int onscreenhelp;               // Shows/hides onscreen help.
+    int multisample;                // Turn off/on multisampling.
+    int ghostboxes;                 // Shows/hides ghost boxes.
+    int reference;                  // reb_particle used as a reference for centering.
+};
+
 struct reb_display_data {
+    struct reb_display_settings s;
     struct reb_simulation* r;
     struct reb_simulation* r_copy;
-    struct reb_vec4d_opengl* particle_data;
+    struct reb_vec4df* particle_data;
     struct reb_orbit_opengl* orbit_data;
     uint64_t N_allocated;
-    double scale;
-    double mouse_scale;
     double mouse_x;
     double mouse_y;
     double retina;
@@ -1112,22 +1138,11 @@ struct reb_display_data {
 #ifdef __EMSCRIPTEN__
     int connection_status;
 #endif
-    int spheres;                    // Switches between point sprite and real spheres.
-    int pause;                      // Pauses visualization, but keep simulation running
-    int wire;                       // Shows/hides orbit wires.
-    int past;                       // Shows/hides past particle positions.
     uint64_t past_last_steps_done;
-    unsigned int past_N;            // Number of past particle positions.
     unsigned int past_N_allocated;
     unsigned int past_current_index;
-    int onscreentext;               // Shows/hides onscreen text.
-    int onscreenhelp;               // Shows/hides onscreen help.
-    int multisample;                // Turn off/on multisampling.
-    int ghostboxes;                 // Shows/hides ghost boxes.
-    int reference;                  // reb_particle used as a reference for centering.
     unsigned int mouse_action;      
     unsigned int key_mods;      
-    struct reb_rotation view;
     unsigned int particle_buffer;
     unsigned int particle_buffer_current;
     unsigned int orbit_buffer;
@@ -1179,6 +1194,21 @@ struct reb_display_data {
 
 };
 
+// Display settings initialization
+DLLEXPORT void reb_simulation_add_display_settings(struct reb_simulation* r);
+
+// Matrix methods
+DLLEXPORT struct reb_mat4df reb_mat4df_identity();
+DLLEXPORT struct reb_mat4df reb_mat4df_scale(struct reb_mat4df m, float x, float y, float z);
+DLLEXPORT void reb_mat4df_print(struct reb_mat4df m);
+DLLEXPORT int reb_mat4df_eq(struct reb_mat4df A, struct reb_mat4df B);
+DLLEXPORT struct reb_vec3df reb_mat4df_get_scale(struct reb_mat4df m);
+DLLEXPORT struct reb_mat4df reb_mat4df_translate(struct reb_mat4df m, float x, float y, float z);
+DLLEXPORT struct reb_mat4df reb_mat4df_multiply(struct reb_mat4df A, struct reb_mat4df B);
+DLLEXPORT struct reb_mat4df reb_rotation_to_mat4df(struct reb_rotation A);
+DLLEXPORT struct reb_mat4df reb_mat4df_ortho(float l, float r, float b, float t, float n, float f);
+
+
 // Declarations and functions needed internally or by python interface only.
 void reb_sigint_handler(int signum);
 
@@ -1208,6 +1238,7 @@ struct reb_binary_field_descriptor {
         REB_FIELD_END = 13,          // Special type to indicate end of blob
         REB_FIELD_NOT_FOUND = 14,    // Special type used to throw error messages
         REB_PARTICLE4 = 15,          // Used for WHFast512
+        REB_POINTER_FIXED_SIZE = 16, // A pointer with a fixed size.
     } dtype;
     char name[1024];
     size_t offset;              // Offset of the storage location relative to the beginning of reb_simulation